# Rework delegate proxy generation approach to improve performance