--- a/src/NSubstitute/Core/ICallFactory.cs
+++ b/src/NSubstitute/Core/ICallFactory.cs
@@ -8,7 +8,6 @@ namespace NSubstitute.Core
     public interface ICallFactory
     {
         ICall Create(MethodInfo methodInfo, object[] arguments, object target, IList<IArgumentSpecification> argumentSpecifications, Func<object> baseMethod);
-        ICall Create(MethodInfo methodInfo, object[] arguments, object target, IList<IArgumentSpecification> argumentSpecifications, IParameterInfo[] parameterInfos);
         ICall Create(MethodInfo methodInfo, object[] getterArgs, object target, IList<IArgumentSpecification> getArgumentSpecifications);
     }
 }