--- a/src/NSubstitute/Core/ICallRouterResolver.cs
+++ b/src/NSubstitute/Core/ICallRouterResolver.cs
@@ -2,7 +2,6 @@ namespace NSubstitute.Core
 {
     public interface ICallRouterResolver
     {
-        void Register(object proxy, ICallRouter callRouter);
         ICallRouter ResolveFor(object substitute);
     }
 }