--- a/src/NSubstitute/Proxies/DelegateProxy/ProxiedDelegateTypeAttribute.cs
+++ b/src/NSubstitute/Proxies/DelegateProxy/ProxiedDelegateTypeAttribute.cs
@@ -1,0 +1,15 @@
+using System;
+
+namespace NSubstitute.Proxies.DelegateProxy
+{
+    [AttributeUsage(AttributeTargets.Method)]
+    public class ProxiedDelegateTypeAttribute : Attribute
+    {
+        public Type DelegateType { get; }
+
+        public ProxiedDelegateTypeAttribute(Type delegateType)
+        {
+            DelegateType = delegateType;
+        }
+    }
+}