Fixes #248.

I've investigated the delegate generation approach and observed a very interesting thing. I noticed, that in fact delegate proxy doesn't differ too  much from the delegate for the member of the regular Castle proxy. Therefore, I reworked the delegate factory to use Castle proxies.

The algorithm is the following:
1. Dynamically generate interface type containing a single method with signature, compatible with the requested delegate.
2. Ask Castle to generate proxy for the interface from the point 1.
3. Generate delegate, bound to the proxy's method and return it.

That allowed to use Castle to generate the method interceptor, providing us with the cool performance.

Test body:
```c#
namespace StaticMocks.Benchmarks
{
    using System;
    using NSubstitute;
    using BenchmarkDotNet.Attributes;
    using BenchmarkDotNet.Attributes.Columns;

    [MinColumn, MaxColumn]
    public class NSubstituteBenchmark
    {
        [Benchmark]
        public object SubstituteForInterface()
        {
            return Substitute.For<IServiceProvider>();
        }

        [Benchmark]
        public object SubstituteForClass()
        {
            return Substitute.For<Type>();
        }

        [Benchmark]
        public object SubstituteForDelegate()
        {
            return Substitute.For<Action>();
        }
    }
}
```

Result: 
```
BenchmarkDotNet=v0.10.12, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.248)
Intel Core i7-6700K CPU 4.00GHz (Skylake), 1 CPU, 8 logical cores and 4 physical cores
Frequency=3914057 Hz, Resolution=255.4894 ns, Timer=TSC
  [Host]     : .NET Framework 4.6.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2633.0
  DefaultJob : .NET Framework 4.6.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2633.0


                 Method |      Mean |     Error |    StdDev |       Min |       Max |
----------------------- |----------:|----------:|----------:|----------:|----------:|
 SubstituteForInterface |  5.165 us | 0.0274 us | 0.0257 us |  5.123 us |  5.211 us |
     SubstituteForClass |  5.399 us | 0.0657 us | 0.0549 us |  5.299 us |  5.487 us |
  SubstituteForDelegate | 10.312 us | 0.0472 us | 0.0394 us | 10.250 us | 10.381 us |
```

As you can see, now the speed is almost comparable.

Notice, I've found that performance is poor if you always generate substitute for the unique delegate types. 95% of the slowdown is caused by the performance of Castle.Proxies and story is exactly the same for the regular proxies as well, so likely no issue here. I apply the caching inside the factory, so I re-use the generated interface type if you request substitute for the same delegate type again.

----

I've found that this approach required me to apply a few tweaks only to make it work (mostly related to how the method is rendered). A good sign that nothing important changed ☺️ 

One more cool benefit that it allowed me to rework the `CallRouterResolver` to no longer store all the generated proxies to later resolve `ICallRouter` for them - very positive change. That also allowed to improve performance significantly, as insertion speed degraded over time due to hash collisions.

As usual, none of the existing tests failed, so I believe that PR doesn't contain breaking changes and my assumption is correct - there is indeed no difference.

Let me know if you see some issues with the new approach I follow.

Thanks for your review 🍷