@zvirja 2018-03-01
@dtchepak I've addressed your concerns. Could you please test my changes on your large codebase, so we can see whether there some issues with it?