@zvirja 2018-03-01
Thanks for the quick merge :) It's cool that solution was so simple 🎉