@jcansdale 2018-05-02
@zvirja This is great! 😄 I'm glad the solution turned out to be simple. Nice to have consistent performance with delegates! 🎉